requirements.txt
langchain-core
langgraph
langchain-openai
langchain-ollama
langchain-chroma
ultralytics
pydantic
python-dotenv
